Adding new ad inventory for SLAM
Adding the new div-gpt-ad numbers and SLAM DFP slot nomenclature to your codes and paths.

I'm not sure how you want to handle the return [ .... ] part, so I didn't take a stab at it. 

We'll be loading a second instance of 728x90 on desktop and a second instance of 320x50 on mobile.